requirements.txt
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
entrypoints==0.4
idna==3.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==7.1.2
jupyter-core==4.9.2
nest-asyncio==1.5.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
tgtg==0.11.1
tornado==6.1
traitlets==4.3.3
urllib3==1.26.9
wcwidth==0.2.5
